linux-secure: FIPS canister build for 6.0.7 secure kernel
- Revert static calls with indirect calls in aesni_intel-glue.c
- Discard ".note.gnu.property" section from canister measurement
- Remove des3_ede tests from crypto_self_test module
- Fix crypto/Makefile warnings
- Enable DES3_IS_NOT_ALLOWED macro in fips_canister_wrapper
- Enable kernel config AES_NI_INTEL as 'y'

Reviewed-on: http://photon-jenkins.eng.vmware.com:8082/18762
Tested-by: Alexey Makhalov <amakhalov@vmware.com>

+++ SPECS/linux/canister_builder/crypto/fips_integrity.h
@@ -1,7 +1,7 @@
 /*
  * FIPS Integrity check for Crypto API
  *
- * Copyright (C) 2020, 2021, VMware, Inc.
+ * Copyright (C) 2020 - 2022 VMware, Inc.
  * Author: Alexey Makhalov <amakhalov@vmware.com>
  *
  */
@@ -23,7 +23,7 @@ struct __attribute__((packed)) relocation {
 	unsigned char section : SECTION_BITS;
 	unsigned char type : TYPE_BITS;
 	unsigned short symbol : SYMBOL_BITS;
-	unsigned short offset;
+	unsigned int offset;
 	int addend;
 };
